Completed excercise 05_sumAll
First created sumAll function with a for loop. Then added the type of
check to make sure non numbers returned an error message. Then added
a check for negative numbers. Last was to switch params 1 and 2 if 1
was higher then 2.

Can refactor to limit amount of nested conditional statements and
repeated for loop.