Completed excercise 06_leapYears
Created a condition to seperate all numbers that are divisible
by four. Then a second set that ensures numbers divisible by four
will be divisible by 400 but not 100 if they're big enough. This is
done without using a conditional to test the size of the number,
but buy ordering the conditional statements in the correct order.
The 400 test first, 100 test second, under 100 test last.